feat: update Dashboard layout component structure
