--- src/components/domain/DashBoard/Layout/Header/index.tsx
@@ -1,3 +1,0 @@
-import Header from './Header';
-
-export default Header;
